requirements.txt
requests==2.25.1
PySocks==1.7.1
